refactoring summary form componenet and adding formerror if something is missing
