+++ apps/financial-aid/web-osk/src/components/index.ts
@@ -18,3 +18,7 @@ export { default as InProgress } from './Status/InProgress'
 export { default as Approved } from './Status/Approved'
 export { default as Rejected } from './Status/Rejected'
 export { default as FileUploadResult } from './FileUploadResult/FileUploadResult'
+export { default as UserInfo } from './Summary/UserInfo'
+export { default as AllFiles } from './Summary/AllFiles'
+export { default as FormInfo } from './Summary/FormInfo'
+export { default as FormComment } from './Summary/FormComment'
